+++ src/fr/adrienbrault/idea/symfony2plugin/SettingsForm.java
@@ -56,6 +56,8 @@ public class SettingsForm implements Configurable {
     private TextFieldWithBrowseButton pathToTranslationRootTextField;
     private JButton buttonHelp;
     private JCheckBox highlightServices;
+    private JCheckBox twigAnnotateTranslation;
+    private JCheckBox phpAnnotateTranslation;
 
     public SettingsForm(@NotNull final Project project) {
         this.project = project;
@@ -112,14 +114,16 @@ public class SettingsForm implements Configurable {
             || !twigAnnotateTemplate.isSelected() == getSettings().twigAnnotateTemplate
             || !twigAnnotateAsset.isSelected() == getSettings().twigAnnotateAsset
             || !twigAnnotateAssetTags.isSelected() == getSettings().twigAnnotateAssetTags
+            || !twigAnnotateTranslation.isSelected() == getSettings().twigAnnotateTranslation
 
             || !phpAnnotateTemplate.isSelected() == getSettings().phpAnnotateTemplate
             || !phpAnnotateService.isSelected() == getSettings().phpAnnotateService
             || !phpAnnotateRoute.isSelected() == getSettings().phpAnnotateRoute
             || !phpAnnotateTemplateAnnotation.isSelected() == getSettings().phpAnnotateTemplateAnnotation
+            || !phpAnnotateTranslation.isSelected() == getSettings().phpAnnotateTranslation
+            || !highlightServices.isSelected() == getSettings().phpHighlightServices
 
             || !yamlAnnotateServiceConfig.isSelected() == getSettings().yamlAnnotateServiceConfig
-            || !highlightServices.isSelected() == getSettings().phpHighlightServices
 
             || !directoryToApp.getText().equals(getSettings().directoryToApp)
             || !directoryToWeb.getText().equals(getSettings().directoryToWeb)
@@ -143,11 +147,13 @@ public class SettingsForm implements Configurable {
         getSettings().twigAnnotateTemplate = twigAnnotateTemplate.isSelected();
         getSettings().twigAnnotateAsset = twigAnnotateAsset.isSelected();
         getSettings().twigAnnotateAssetTags = twigAnnotateAssetTags.isSelected();
+        getSettings().twigAnnotateTranslation = twigAnnotateTranslation.isSelected();
 
         getSettings().phpAnnotateTemplate = phpAnnotateTemplate.isSelected();
         getSettings().phpAnnotateService = phpAnnotateService.isSelected();
         getSettings().phpAnnotateRoute = phpAnnotateRoute.isSelected();
         getSettings().phpAnnotateTemplateAnnotation = phpAnnotateTemplateAnnotation.isSelected();
+        getSettings().phpAnnotateTranslation = phpAnnotateTranslation.isSelected();
         getSettings().phpHighlightServices = highlightServices.isSelected();
 
         getSettings().yamlAnnotateServiceConfig = yamlAnnotateServiceConfig.isSelected();
@@ -185,11 +191,13 @@ public class SettingsForm implements Configurable {
         twigAnnotateTemplate.setSelected(getSettings().twigAnnotateTemplate);
         twigAnnotateAsset.setSelected(getSettings().twigAnnotateAsset);
         twigAnnotateAssetTags.setSelected(getSettings().twigAnnotateAssetTags);
+        twigAnnotateTranslation.setSelected(getSettings().twigAnnotateTranslation);
 
         phpAnnotateTemplate.setSelected(getSettings().phpAnnotateTemplate);
         phpAnnotateService.setSelected(getSettings().phpAnnotateService);
         phpAnnotateRoute.setSelected(getSettings().phpAnnotateRoute);
         phpAnnotateTemplateAnnotation.setSelected(getSettings().phpAnnotateTemplateAnnotation);
+        phpAnnotateTranslation.setSelected(getSettings().phpAnnotateTranslation);
         highlightServices.setSelected(getSettings().phpHighlightServices);
 
         yamlAnnotateServiceConfig.setSelected(getSettings().yamlAnnotateServiceConfig);